refactor(cli): keep commands.rs as thin clap dispatch
Move logging, config loading, and subcommand clap types into
commands/ modules so commands.rs only defines the top-level
parser and forwards to module run functions.

Closes #1954
