+++ crates/harness-cli/src/commands/execpolicy.rs
@@ -1,5 +1,35 @@
-use super::{configured_rule_engine, ExecPolicyCommand};
+use clap::Subcommand;
 use harness_core::config::HarnessConfig;
+use std::path::PathBuf;
+
+use super::rule::configured_rule_engine;
+
+#[derive(Subcommand)]
+pub enum ExecPolicyCommand {
+    /// Check a command against Starlark policy rules
+    Check {
+        /// Paths to policy files (repeatable). Falls back to `rules.exec_policy_paths`.
+        #[arg(short = 'r', long = "rules", value_name = "PATH")]
+        rules: Vec<PathBuf>,
+        /// Optional requirements.toml path. Falls back to `rules.requirements_path` when omitted.
+        #[arg(long, value_name = "PATH")]
+        requirements: Option<PathBuf>,
+        /// Resolve absolute executables against basename rules.
+        #[arg(long)]
+        resolve_host_executables: bool,
+        /// Pretty-print JSON output.
+        #[arg(long)]
+        pretty: bool,
+        /// Command tokens to evaluate.
+        #[arg(
+            value_name = "COMMAND",
+            required = true,
+            trailing_var_arg = true,
+            allow_hyphen_values = true
+        )]
+        command: Vec<String>,
+    },
+}
 
 pub fn run(cmd: ExecPolicyCommand, config: &HarnessConfig) -> anyhow::Result<()> {
     match cmd {
