+++ crates/harness-cli/src/commands/runtime_log_tests.rs
@@ -1,5 +1,10 @@
-use super::*;
-use chrono::TimeZone;
+use super::logging::{
+    prepare_logging, purge_stale_runtime_logs, purge_stale_runtime_logs_with, runtime_log_path,
+};
+use super::Command;
+use chrono::{TimeZone, Utc};
+use std::fs;
+use std::io;
 
 #[test]
 fn prepare_logging_creates_runtime_log_for_serve() {
